published new versions (patch) of packages: @bemoje/repoman@0.0.39

+++ packages/repoman/src/lib/wipeNodeModules.ts
@@ -3,30 +3,64 @@ import fs from 'fs'
 import path from 'path'
 import { getPackages } from './getPackages'
 
-export function wipeNodeModules(args: string[]) {
-  getPackages().forEach(({ name, rootdir }) => {
-    const dir = path.join(rootdir, 'node_modules')
-    const plock = path.join(rootdir, 'package-lock.json')
-    if (fs.existsSync(dir)) {
-      deleteDirectorySafeSync(dir)
-      console.log(`Deleted ${name}/node_modules/`)
-    }
-    if (args.includes('lock') && fs.existsSync(plock)) {
-      fs.rmSync(plock, { force: true })
-      console.log(`Deleted ${name}/package-lock.json`)
-    }
-  })
+export function wipeNodeModules(
+  packages: string[],
+  options: { scope?: string; packageLock?: boolean; root?: boolean } = {}
+) {
+  console.log({ packages, options })
+  return
+  if (options.scope) {
+    const scope = '@' + options.scope
+    getPackages(packages).forEach(({ name, rootdir }) => {
+      const dir = path.join(rootdir, 'node_modules', scope)
+      const plock = path.join(rootdir, 'package-lock.json')
+      if (fs.existsSync(dir)) {
+        deleteDirectorySafeSync(dir)
+        console.log(`Deleted ${name}/node_modules/` + options.scope)
+      }
 
-  if (args.includes('all')) {
-    const dir = path.join(process.cwd(), 'node_modules')
-    const plock = path.join(process.cwd(), 'package-lock.json')
-    if (fs.existsSync(dir)) {
-      deleteDirectorySafeSync(dir)
-      console.log(`Deleted ${'tsmono'}/node_modules`)
+      if (options.packageLock && fs.existsSync(plock)) {
+        fs.rmSync(plock, { force: true })
+        console.log(`Deleted ${name}/package-lock.json`)
+      }
+    })
+    if (options.root) {
+      const dir = path.join(process.cwd(), 'node_modules', scope)
+      const plock = path.join(process.cwd(), 'package-lock.json')
+      if (fs.existsSync(dir)) {
+        deleteDirectorySafeSync(dir)
+        console.log(`Deleted ${'tsmono'}/node_modules/` + scope)
+      }
+      if (options.packageLock && fs.existsSync(plock)) {
+        fs.rmSync(plock, { force: true })
+        console.log(`Deleted ${'tsmono'}/package-lock.json`)
+      }
     }
-    if (args.includes('lock') && fs.existsSync(plock)) {
-      fs.rmSync(plock, { force: true })
-      console.log(`Deleted ${'tsmono'}/package-lock.json`)
+  } else {
+    getPackages(packages).forEach(({ name, rootdir }) => {
+      const dir = path.join(rootdir, 'node_modules')
+      const plock = path.join(rootdir, 'package-lock.json')
+      if (fs.existsSync(dir)) {
+        deleteDirectorySafeSync(dir)
+        console.log(`Deleted ${name}/node_modules/`)
+      }
+      if (options.packageLock && fs.existsSync(plock)) {
+        fs.rmSync(plock, { force: true })
+        console.log(`Deleted ${name}/package-lock.json`)
+      }
+    })
+
+    if (options.root) {
+      const dir = path.join(process.cwd(), 'node_modules')
+      const plock = path.join(process.cwd(), 'package-lock.json')
+      if (fs.existsSync(dir)) {
+        deleteDirectorySafeSync(dir)
+        console.log(`Deleted ${'tsmono'}/node_modules`)
+      }
+      if (options.packageLock && fs.existsSync(plock)) {
+        fs.rmSync(plock, { force: true })
+        console.log(`Deleted ${'tsmono'}/package-lock.json`)
+      }
     }
   }
 }
